Implement own DownloaderWorker, allow user to choose
+ fix update check notifications using wrong notification channel

+++ app/src/main/java/ua/gardenapple/itchupdater/database/installation/GameInstallation.kt
@@ -7,7 +7,7 @@ import ua.gardenapple.itchupdater.database.game.Game
 data class GameInstallation(
     @Embedded val game: Game,
     val status: Int,
-    val downloadOrInstallId: Int?,
+    val downloadOrInstallId: Long?,
     val packageName: String?,
     val installId: Int,
     val uploadId: Int,
