+++ app/src/main/java/ua/gardenapple/itchupdater/installer/InstallNotificationBroadcastReceiver.kt
@@ -22,7 +22,7 @@ class InstallNotificationBroadcastReceiver : BroadcastReceiver() {
         Log.d(LOGGING_TAG, "onReceive")
         val extras = intent.extras!!
 
-        val downloadId = extras.getInt(EXTRA_DOWNLOAD_ID)
+        val downloadId = extras.getLong(EXTRA_DOWNLOAD_ID)
 
         runBlocking {
             val apkFile = File(intent.data!!.path!!)
